backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
pymongo==4.6.1
python-dotenv==1.0.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
joblib==1.3.2
Pillow==10.1.0
pytesseract==0.3.10
reportlab==4.0.7
matplotlib==3.8.2
seaborn==0.13.0
bcrypt==4.1.2
gunicorn==21.2.0
transformers==4.36.2
torch==2.1.2
